+++ mypy.ini
@@ -6,4 +6,4 @@ follow_imports = silent
 strict = True
 warn_redundant_casts = False
 warn_unused_ignores = False
-packages = nionswift_plugin.DM_IO, nionswift_plugin.TIFF_IO, DM_IO
+packages = nionswift_plugin.DM_IO, nionswift_plugin.TIFF_IO, nion/io/DM_IO
