+++ pyproject.toml
@@ -33,15 +33,15 @@ include = [
     "nionswift_plugin",
     "nionswift_plugin.TIFF_IO",
     "nionswift_plugin.TIFF_IO.test",
-    "DM_IO",
-    "DM_IO.test"
+    "nion/io/DM_IO",
+    "nion/io/DM_IO.test"
 ]
 
 [tool.setuptools.package-data]
-"DM_IO.test" = ["resources/*"]
+"nion/io/DM_IO.test" = ["resources/*"]
 
 [tool.pytest.ini_options]
 testpaths = [
     "nionswift_plugin/TIFF_IO",
-    "DM_IO"
+    "nion/io/DM_IO"
 ]
